+++ sdk/dotnet/Binaryauthorization/AttestorIamBinding.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi.Serialization;
 
-namespace Pulumi.Gcp.Binaryauthorization
+namespace Pulumi.Gcp.BinaryAuthorization
 {
     /// <summary>
     /// &gt; This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/r/binary_authorization_attestor_iam_binding.html.markdown.