+++ sdk/dotnet/Runtimeconfig/Config.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Pulumi.Serialization;
 
-namespace Pulumi.Gcp.Runtimeconfig
+namespace Pulumi.Gcp.RuntimeConfig
 {
     /// <summary>
     /// Manages a RuntimeConfig resource in Google Cloud. For more information, see the