+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Namespace names in .NET SDK are adjusted to CamelCase
+* Namespace names in .NET SDK are adjusted to PascalCase
 ([#259](https://github.com/pulumi/pulumi-gcp/pull/259)).
 
 ---